게시판 수정

+++ final-pjt-front/src/components/CommentListItem.vue
@@ -1,41 +1,41 @@
 <template>
-  <!-- <div v-if="comment.article.pk === "> -->
   <div v-if="Number(route.params.id) === comment.article.pk">
-    <!-- <p>글 pk : {{ route.params.id }}</p>
-    <p>댓글 글 pk : {{ comment.article.pk }}</p> -->
-    <!-- <p>{{ comment }}</p> -->
     <p>작성자 : {{ comment.user.username }}</p>
-    <!-- <p>{{ comment.id }}</p> -->
-
-
-
-    <div v-if="showContent">
-      <p>내용 : {{ comment.content }}</p>
-      <button @click="clickUpdate">수정</button>
-      <button @click="deleteComment(comment.id)">삭제</button>
+    <div v-if="comment.user.email === loginUser && loginUser !== null">
+      <div v-if="showContent">
+        <p>내용 : {{ comment.content }}</p>
+        <p v-if="comment.created_at !== comment.updated_at">
+        (수정됨)
+        </p>
+        <button @click="clickUpdate">수정</button>
+        <button @click="deleteComment(comment.id)">삭제</button>
+      </div>
+      <div v-else>
+        <form @submit.prevent="updateComment(comment.id)">
+          <div>
+            <label for="content">내용:</label>
+            <textarea v-model.trim="content" id="content"></textarea>
+          </div>
+          <input type="submit" value="수정">
+        </form>
+      </div>
     </div>
     <div v-else>
-      <form @submit.prevent="updateComment(comment.id)">
-        <div>
-          <label for="content">내용:</label>
-          <textarea v-model.trim="content" id="content"></textarea>
-        </div>
-        <input type="submit" value="수정">
-      </form>
-      <!-- <UpdateComment/> -->
+      <p>내용 : {{ comment.content }}</p>
+      <p v-if="comment.created_at !== comment.updated_at">
+      (수정됨)
+      </p>
     </div>
-
     <hr>
   </div>
 </template>
 
 <script setup>
-// import UpdateComment from './UpdateComment.vue';
 import { useRouter, useRoute } from 'vue-router';
 import { useCommunityStore } from '@/stores/community';
 import { useUserStore } from '@/stores/user'
 import axios from 'axios';
-import { ref } from 'vue'
+import { ref, onMounted } from 'vue'
 
 
 const API_URL = import.meta.env.VITE_API_URL
@@ -44,12 +44,19 @@ const API_URL = import.meta.env.VITE_API_URL
 const route = useRoute()
 const router = useRouter()
 const store = useCommunityStore()
-const content = ref(null)
 const userStore = useUserStore()
 const showContent = ref(true)
+const loginUser = ref(null)
 
+const props = defineProps({
+  comment: Object,
+})
+
+const comment = ref(props.comment)
+const content = ref(comment.value.content)
 
 const updateComment = function(commentId) {
+  if (comment.value.user.email === loginUser.value && loginUser.value !== null) {
     axios({
       method: 'PUT',
       url: `${API_URL}/community/comments/${commentId}/`,
@@ -61,7 +68,6 @@ const updateComment = function(commentId) {
       },
     })
     .then((res) => {
-        // router
         console.log(res)
         showContent.value != showContent.value
         router.go()
@@ -69,30 +75,50 @@ const updateComment = function(commentId) {
     .catch((err) => {
       console.log(err)
     })
+  } else {
+    alert('수정할 수 없습니다.')
+  }
 }
 
 const deleteComment = function(commentId) {
-  axios({
-    method: 'DELETE',
-    url: `${store.API_URL}/community/comments/${commentId}/`
-  })
-  .then(() => {
-    router.go()
-    // alert('댓글이 삭제되었습니다.')
-  })
-  .catch(() => {
-    console.log(err)
-  })
+  if (comment.value.user.email === loginUser.value && loginUser.value !== null) {
+    console.log(1)
+    axios({
+      method: 'DELETE',
+      url: `${store.API_URL}/community/comments/${commentId}/`
+    })
+    .then(() => {
+      router.go()
+      // alert('댓글이 삭제되었습니다.')
+    })
+    .catch(() => {
+      console.log(err)
+    })
+  } else {
+    alert('삭제할 수 없습니다.')
+  }
 }
 
 const clickUpdate = function() {
   showContent.value = !showContent.value
 }
 
-// const updateComment = function(commentId)
+onMounted(() => {
+  axios({
+    method: 'GET',
+    url: `${store.API_URL}/accounts/user/`,
+    headers: {
+      Authorization: `Token ${userStore.token}`
+    }
+  })
+  .then((res) => {
+    loginUser.value = res.data.email
 
-
-defineProps({
-  comment: Object,
+  })
+  .catch((err) => {
+    console.log(err)
+  })
 })
+
+
 </script>
